refactor utils again and use lodash.has to be consistent with set, get, unset
