+++ packages/utils/src/decorators.ts
@@ -1,9 +1,3 @@
-const { NODE_ENV = 'production' } = process.env;
-
-export const isProd = NODE_ENV === 'production';
-export const isTest = NODE_ENV === 'test';
-export const isDev = NODE_ENV === 'development';
-
 /** A decorator for locking down a method */
 export function locked() {
     return function _locked(
