+++ packages/utils/src/index.ts
@@ -2,19 +2,24 @@ import STATUS_CODES from './status-codes';
 
 export * from './arrays';
 export * from './big-map';
+export * from './booleans';
+export * from './buffers';
 export * from './collector';
+export * from './core';
 export * from './entities';
 export * from './dates';
+export * from './decorators';
+export * from './empty';
+export * from './env';
 export * from './errors';
 export * from './functions';
 export * from './geo';
 export * from './interfaces';
+export * from './json';
 export * from './logger';
-export * from './misc';
 export * from './numbers';
 export * from './objects';
 export * from './promises';
 export * from './queue';
 export * from './strings';
-export * from './utils';
 export { STATUS_CODES };
