+++ packages/utils/src/objects.ts
@@ -1,26 +1,12 @@
-import set from 'lodash.set';
-import get from 'lodash.get';
-import unset from 'lodash.unset';
-import cloneDeep from 'lodash.clonedeep';
-import isPlainObject from 'is-plain-object';
 import { WithoutNil, FilteredResult } from './interfaces';
-import { isBooleanLike } from './utils';
+import { isBooleanLike } from './booleans';
+import { get, isPlainObject } from './core';
 
 export function getFirstValue<T>(input: { [key: string]: T }): T | undefined {
     return Object.values(input)[0];
 }
 export function getFirstKey<T>(input: T): (keyof T) | undefined {
     return Object.keys(input)[0] as keyof T;
-}
-
-/** Check in input has a key */
-export function has(data: object|undefined, key: string|number|symbol): boolean {
-    if (data == null || typeof data !== 'object') return false;
-    if (data instanceof Set || data instanceof Map) {
-        if (key in data) return true;
-        return data.has(key);
-    }
-    return key in data;
 }
 
 /**
@@ -94,18 +80,6 @@ export function withoutNil<T extends object>(input: T): WithoutNil<T> {
     }
 
     return result as WithoutNil<T>;
-}
-
-/**
- * Similar to is-plain-object but works better when clone deeping a DataEntity
-*/
-export function isSimpleObject(input: any): input is object {
-    if (input == null) return false;
-    if (Buffer.isBuffer(input)) return false;
-    if (Array.isArray(input)) return false;
-    if (input instanceof Set) return false;
-    if (input instanceof Map) return false;
-    return typeof input === 'object';
 }
 
 /**
@@ -183,12 +157,3 @@ export function getField<T, P extends keyof T, V>(
     }
     return result || defaultVal;
 }
-
-// export a few dependencies
-export {
-    isPlainObject,
-    cloneDeep,
-    get,
-    set,
-    unset
-};
