+++ packages/utils/src/promises.ts
@@ -1,4 +1,4 @@
-import { isEmpty } from './utils';
+import { isEmpty } from './empty';
 import { debugLogger } from './logger';
 import { toHumanTime, trackTimeout } from './dates';
 import {
