+++ types/shallow-clone/index.d.ts
@@ -1,0 +1,3 @@
+/** Declaration file generated by dts-gen */
+
+export default function clone(input: any): any;
